refactor(EventBadge): EventBadge가 직접 배지를 만들도록 EventBadgeGenerator클래스 제거
