+++ src/main/java/christmas/domain/entity/EventBadge.java
@@ -6,6 +6,9 @@ public enum EventBadge {
     STAR("별"),
     TREE("트리"),
     SANTA("산타");
+    private static final int SANTA_THRESHOLD = 20000;
+    private static final int TREE_THRESHOLD = 10000;
+    private static final int STAR_THRESHOLD = 5000;
 
     private final String name;
 
@@ -15,6 +18,20 @@ public enum EventBadge {
 
     public String getName() {
         return name;
+    }
+
+    public static EventBadge createBadge(DiscountRecord discountRecord) {
+        int totalBenefitAmount = discountRecord.calculateTotalBenefitAmount();
+        if (totalBenefitAmount >= SANTA_THRESHOLD) {
+            return EventBadge.SANTA;
+        }
+        if (totalBenefitAmount >= TREE_THRESHOLD) {
+            return EventBadge.TREE;
+        }
+        if (totalBenefitAmount >= STAR_THRESHOLD) {
+            return EventBadge.STAR;
+        }
+        return EventBadge.NONE;
     }
 
 }
