+++ src/test/java/christmas/domain/entity/EventBadgeTest.java
@@ -1,9 +1,13 @@
 package christmas.domain.entity;
 
+import christmas.domain.entity.subclass.BadgeCreator;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.CsvSource;
 
+import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.*;
 
 class EventBadgeTest {
@@ -23,5 +27,22 @@ class EventBadgeTest {
         }
     }
 
+    @Nested
+    class createBadge {
+        @ParameterizedTest
+        @DisplayName("createBadge 메소드 테스트")
+        @CsvSource(value = {
+                "100:없음",
+                "3241:없음",
+                "5312:별",
+                "11123:트리",
+                "23513:산타",
+                "75432:산타"
+        }, delimiter = ':')
+        void 총_혜택_금액에_따른_이번트_배지_테스트(int totalBenefitAmount, String expected) {
+            BadgeCreator badgeCreator = new BadgeCreator();
+            assertThat(badgeCreator.createBadge(totalBenefitAmount).getName()).isEqualTo(expected);
+        }
+    }
 
 }
